move all raw commands in different file
